pyproject.toml
[project]
name = "persbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiohttp>=3.8.0",
    "chzzkpy>=1.0.0",
    "discord-py>=2.3.0",
    "ddgs>=8.0.0",
    "google-genai>=1.50.0",
    "openai>=1.59.6",
    "numpy>=2.2.6",
    "psutil>=5.9.0",
    "python-dotenv>=1.0.0",
    "sqlalchemy>=2.0.44",
    "aiofiles>=25.1.0",
    "beautifulsoup4>=4.12.0",
    "pillow>=10.0.0",
    "pytz>=2024.1",
    "pytest>=9.0.0",
    "pytest-asyncio>=1.3.0",
    "pytest-mock>=3.15.1",
    "pytest-cov>=7.0.0",
    "dpytest>=0.7.0",
]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "W", "I"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
warn_return_any = true
warn_unused_configs = true
[[tool.mypy.overrides]]
module = "persbot.services.model_usage_service"
ignore_missing_imports = false
[[tool.mypy.overrides]]
module = "persbot.services.prompt_service"
ignore_missing_imports = false

[dependency-groups]
dev = [
    "black>=26.1.0",
    "mypy>=1.19.1",
    "ruff>=0.15.0",
]
